+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Added
 
 - add graph modification when less than graph returns are specified
 - add "__name__" attribute to handler instance
+- ``signature_modifier`` can modify function with "**kwargs"
 
 [0.4.0] - 2022-10-3
 ------------------------